requirements.txt
# IMPORTANT: Install PyTorch separately for your CUDA version.
# For CUDA 12.1 (works on CUDA 12.3 systems):
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

opencv-python
pillow
nevergrad
ultralytics
scikit-learn
numpy
tqdm
requests
click
matplotlib
pyyaml
yaml